pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "credit-portfolio"
version = "1.0.0"
description = "Three-layer systematic credit portfolio construction: HMM regime detection, Black-Litterman, CVXPY optimization"
requires-python = ">=3.10"
dependencies = [
    "cvxpy>=1.4",
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "requests>=2.28",
    "anthropic>=0.25",
    "hmmlearn>=0.3",
    "prophet>=1.1",
    "pandas-datareader>=0.10",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
    "shap>=0.43",
    "streamlit>=1.30",
    "plotly>=5.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
credit-portfolio = "credit_portfolio.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
